wwwwp

+++ SpringBoot/projectA/src/main/java/com/projectA/VO/Al_inputVO.java
@@ -1,5 +1,7 @@
 package com.projectA.VO;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,10 +11,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Al_inputVO {
     private Long inputIdx;
-    private int tensileStrength;
-    private int yieldStrength;
-    private int hardness;
-    private int elongation;
+    private BigDecimal tensileStrength;
+    private BigDecimal yieldStrength;
+    private BigDecimal hardness;
+    private BigDecimal elongation;
     private char firstSolution;
     private char secondSolution;
     private char aging;
